middleware/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-dotenv
requests
orjson